source code cleanup, step 3: eclipse code cleanup - removing unused casts

+++ makumba/classes/org/makumba/importer/ObjectImporter.java
@@ -330,7 +330,7 @@ public class ObjectImporter {
         }
         Iterator<Integer> f = dd.getFieldDefinition(fieldName).getValues().iterator();
         for (String v : dd.getFieldDefinition(fieldName).getNames()) {
-            Integer i = (Integer) f.next();
+            Integer i = f.next();
             if (v.equals(s)) {
                 return i;
             }
